a draft for middle layer
