+++ src/lava/magma/runtime/message_infrastructure/message_infrastructure/csrc/shmem_channel.cc
@@ -27,8 +27,8 @@ ShmemChannel::ShmemChannel(const std::string &src_name,
                            const size_t &nbytes) {
   shm_ = GetSharedMemManager().AllocChannelSharedMemory(nbytes);
 
-  send_port_ = std::make_shared<ShmemSendPort>(src_name, shm_, size, nbytes);
-  recv_port_ = std::make_shared<ShmemRecvPort>(dst_name, shm_, size, nbytes);
+  send_port_ = std::make_shared<ShmemSendPort>(src_name, shm_, size, nbytes + offsetof(MetaData, mdata));
+  recv_port_ = std::make_shared<ShmemRecvPort>(dst_name, shm_, size, nbytes + offsetof(MetaData, mdata));
 }
 
 AbstractSendPortPtr ShmemChannel::GetSendPort() {
